add elm-with-file
This is similar to elx-with-file but instead of inserting a file from the
livefs into a buffer a blob stored in a git repository is inserted.

Unlike elx-with-file this always creates a new tempory buffer even if the
blob is already inserted into the current buffer.

This function replaces epkg-with-file but is much better.